+++ History.md
@@ -16,6 +16,7 @@
 * Adds property overriding so more coarse properties override more granular ones.
 * Adds reducing non-adjacent selectors.
 * Adds `--skip-advanced`/`noAdvanced` switch to disable advanced optimizations.
+* Adds reducing non-adjacent selectors when overridden by more complex selectors.
 
 1.1.7 / 2013-10-28
 ==================